+++ src/Sovren.SDK/Models/Resume/Skills/ResumeSkill.cs
@@ -4,6 +4,7 @@
 // within the Terms of Service pertaining to the Sovren SaaS products.
 
 using Sovren.Models.Skills;
+using System;
 using System.Collections.Generic;
 
 namespace Sovren.Models.Resume.Skills
@@ -24,7 +25,7 @@ namespace Sovren.Models.Resume.Skills
         /// <summary>
         /// If this skill has any varitaions, this describes the most recent date any of the varitaions were used
         /// </summary>
-        public SovrenDate ChildrenLastUsed { get; set; }
+        public SovrenPrimitive<DateTime> ChildrenLastUsed { get; set; }
     }
 
     /// <inheritdoc/>
@@ -38,6 +39,6 @@ namespace Sovren.Models.Resume.Skills
         /// <summary>
         /// Describes the date the candidate last used the skill (derived from position dates)
         /// </summary>
-        public SovrenDate LastUsed { get; set; }
+        public SovrenPrimitive<DateTime> LastUsed { get; set; }
     }
 }
